pyproject.toml
[project]
name = "agentanalytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==5.1.0",
    "anthropic>=0.77.0",
    "crewai>=0.108.0",
    "duckdb>=1.4.4",
    "ipykernel>=7.1.0",
    "pandas<3",
    "plotly==5.24.1",
    "python-dotenv>=1.2.1",
    "streamlit==1.28.1",
    "tabulate>=0.10.0",
]

[tool.setuptools.packages.find]
where = ["ai_analytics_agent/src"]
